Revamp `exam 3`
