+++ src/main/kotlin/com/dicoding/exam/exam3/Exam3.kt
@@ -1,0 +1,6 @@
+package com.dicoding.exam.exam3
+
+// TODO
+fun <T> checkType(args: T): String {
+    return ""
+}
